moved getTravel to mapConstants.ts
